Fix: API 연동 및 버그 수정 (#49)
- 사용자 정보 CRUD API 연동 구현
 - 사용자 정보 조회/수정/삭제 기능 구현
 - 비밀번호 초기화 기능 구현

- API 요청 버그 수정
 - URL 템플릿 리터럴 문법 오류 수정
 - fetch 요청 URL 문자열 연결 방식으로 변경
 - API 엔드포인트 호출 정상화

- 기타 개선사항
 - 사용자 정보 수정 모달 구현
 - 입력 폼 유효성 검사 추가
 - 응답 처리 및 에러 핸들링 구현
