--- admin/src/main/webapp/views/index.jsp
@@ -1,43 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
-<%@ taglib prefix="sec" uri="http://www.springframework.org/security/tags" %>
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>사내 업무 관리 시스템</title>
-    <!-- Bootstrap CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
-    <!-- Custom CSS -->
-    <link rel="stylesheet" href="/css/style.css">
-    <!-- Font Awesome -->
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" rel="stylesheet">
-</head>
-<body>
-<div class="wrapper">
-    <!-- Sidebar -->
-    <jsp:include page="layout/sidebar.jsp" />
-
-    <div class="main-content">
-        <!-- Header -->
-        <jsp:include page="layout/header.jsp" />
-
-        <!-- Content Area -->
-        <div class="content-wrapper">
-            <div class="container-fluid">
-                <jsp:include page="${contentPage}" />
-            </div>
-        </div>
-
-        <!-- Footer -->
-        <jsp:include page="layout/footer.jsp" />
-    </div>
-</div>
-
-<!-- Bootstrap JS -->
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
-<!-- Custom JS -->
-<script src="/js/script.js"></script>
-</body>
-</html>
