--- admin/src/main/webapp/views/layout/footer.jsp
@@ -1,11 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-
-<footer class="footer">
-    <div class="container-fluid">
-        <div class="row">
-            <div class="col-12">
-                <p class="text-center mb-0">&copy; 2024 사내 업무 관리 시스템. All rights reserved.</p>
-            </div>
-        </div>
-    </div>
-</footer>
